requirements.txt
pymupdf
pytesseract
pandas
pillow
python-docx
openpyxl
